+++ R/s3HTTP.R
@@ -96,7 +96,9 @@ function(verb = "GET",
                query_args = query,
                canonical_headers = canonical_headers,
                request_body = request_body,
-               key = key, secret = secret, session_token = session_token)
+               key = key, 
+               secret = secret, 
+               session_token = session_token)
         headers[["x-amz-date"]] <- d_timestamp
         headers[["x-amz-content-sha256"]] <- Sig$BodyHash
         if (!is.null(session_token) && session_token != "") {
@@ -216,7 +218,12 @@ function(bucketname,
             url_style <- "virtual"
         }
         # handle region
-        if (region == "us-east-1") {
+        if (region %in% c("us-east-1", "")) {
+            if (region == "") {
+                if (isTRUE(verbose)) {
+                    message("Option 'region' is missing, so 'us-east-1' assumed.")
+                }
+            }
             # handle dualstack
             if (isTRUE(dualstack)) {
                 # handle accelerate